refactor(api-user): rename users domain to singular
