+++ libs/api/user/jest.config.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
 export default {
-	displayName: 'api-users',
+	displayName: 'api-user',
 	preset: '../../../jest.preset.js',
 	testEnvironment: 'node',
 	transform: {
 		'^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
 	},
 	moduleFileExtensions: ['ts', 'js', 'html'],
-	coverageDirectory: '../../../coverage/libs/api/users',
+	coverageDirectory: '../../../coverage/libs/api/user',
 };
