+++ libs/api/user/src/lib/core/event/user-deactivated.event.ts
