+++ libs/api/user/src/lib/core/exception/insufficient-permission.exception.ts
