+++ libs/api/user/src/lib/core/exception/user-not-found.exception.ts
