+++ libs/api/user/src/lib/infra/user.view-model.ts
