+++ tsconfig.base.json
@@ -20,7 +20,7 @@
 			"@kordis/api/organization": ["libs/api/organization/src/index.ts"],
 			"@kordis/api/shared": ["libs/api/shared/src/index.ts"],
 			"@kordis/api/test-helpers": ["libs/api/test-helpers/src/index.ts"],
-			"@kordis/api/users": ["libs/api/users/src/index.ts"],
+			"@kordis/api/user": ["libs/api/user/src/index.ts"],
 			"@kordis/shared/auth": ["libs/shared/auth/src/index.ts"],
 			"@kordis/shared/test-helpers": ["libs/shared/test-helpers/src/index.ts"],
 			"@kordis/spa/auth": ["libs/spa/auth/src/index.ts"],
